More warning fixes

+++ src/Fl_File_Icon2.cxx
@@ -264,7 +264,7 @@ Fl_File_Icon::load_fti(const char *fti)	// I - File to read from
     else if (strcmp(command, "bgnoutlinepolygon") == 0)
     {
       add(OUTLINEPOLYGON);
-      outline = add(0) - data_;
+      outline = static_cast<int>(add(0) - data_);
       add(0);
     }
     else if (strcmp(command, "endoutlinepolygon") == 0 && outline)
@@ -479,15 +479,15 @@ int Fl_File_Icon::load_image(const char *ifile)	// I - File to read from
             case 3 :
 		val[0] = lineptr[0];
 		val[1] = '\0';
-		red = 255 * strtol(val, NULL, 16) / 15;
+		red = static_cast<int>(255 * strtol(val, NULL, 16) / 15);
 
 		val[0] = lineptr[1];
 		val[1] = '\0';
-		green = 255 * strtol(val, NULL, 16) / 15;
+		green = static_cast<int>(255 * strtol(val, NULL, 16) / 15);
 
 		val[0] = lineptr[2];
 		val[1] = '\0';
-		blue = 255 * strtol(val, NULL, 16) / 15;
+		blue = static_cast<int>(255 * strtol(val, NULL, 16) / 15);
 		break;
 
             case 6 :
@@ -498,17 +498,17 @@ int Fl_File_Icon::load_image(const char *ifile)	// I - File to read from
 		val[0] = lineptr[0];
 		val[1] = lineptr[1];
 		val[2] = '\0';
-		red = strtol(val, NULL, 16);
+		red = static_cast<int>(strtol(val, NULL, 16));
 
 		val[0] = lineptr[j + 0];
 		val[1] = lineptr[j + 1];
 		val[2] = '\0';
-		green = strtol(val, NULL, 16);
+		green = static_cast<int>(strtol(val, NULL, 16));
 
 		val[0] = lineptr[2 * j + 0];
 		val[1] = lineptr[2 * j + 1];
 		val[2] = '\0';
-		blue = strtol(val, NULL, 16);
+		blue = static_cast<int>(strtol(val, NULL, 16));
 		break;
 	  }
